requirements.txt
Django>=3.2.12,<4.1
djangorestframework>=3.13.1,<3.14
djangorestframework-simplejwt>=5.1.0,<5.2.0
psycopg2>=2.9.3,<2.10
uWSGI>=2.0.20,<2.1
Pillow>=9.0.1,<9.1
flake8>=4.0.1,<4.1
django-filter>=21.1,<22.1 
drf-spectacular>=0.22.0,<0.25.0
drf-spectacular-sidecar>=2022.4.1,<2022.5.1
